# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/{id}": {
      "get": {
        "operationId": "find",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "put": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
...
    "/api/role/{id}": {
      "get": {
        "operationId": "find",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        },
        "tags": [
          "role"
        ]
      },
      "put": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        },
        "tags": [
          "role"
        ]
      },
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "role"
        ]
      }
    },
...
    "/api/user-role/{id}": {
      "get": {
        "operationId": "find",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRole"
                }
              }
            }
          }
        },
        "tags": [
          "user-role"
        ]
      },
      "put": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRole"
                }
              }
            }
          }
        },
        "tags": [
          "user-role"
        ]
      },
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "user-role"
        ]
      }
    },
    "/api/user-role": {
      "get": {
        "operationId": "query",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "where",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "count",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRoleQueryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "user-role"
        ]
      },
      "put": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRole"
                }
              }
            }
          }
        },
        "tags": [
          "user-role"
        ]
      }
    },
...
  },
  "info": {
    "title": "Quasar Nest example",
    "description": "The cats API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Session": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "sessionId",
          "userId",
          "isDeleted",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "object"
          },
          "salt": {
            "type": "object"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRole"
            }
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Session"
            }
          }
        },
        "required": [
          "userId",
          "firstName",
          "lastName",
          "email",
          "password",
          "salt",
          "isDeleted",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Role": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRole"
            }
          }
        },
        "required": [
          "roleId",
          "name",
          "isDeleted",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "UserRole": {
        "type": "object",
        "properties": {
          "userRoleId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "roleId": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "userRoleId",
          "userId",
          "roleId",
          "isDeleted",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
      "ConnectUserRoleDto": {
        "type": "object",
        "properties": {
          "userRoleId": {
            "type": "string"
          }
        },
        "required": [
          "userRoleId"
        ]
      },
...
      "ConnectSessionDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          }
        },
        "required": [
          "sessionId"
        ]
      },
...
      "UpdateUserRolesRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectUserRoleDto"
            }
          }
        },
        "required": [
          "connect"
        ]
      },
      "UpdateUserSessionsRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectSessionDto"
            }
          }
        },
        "required": [
          "connect"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "object"
          },
          "salt": {
            "type": "object"
          },
          "roles": {
            "$ref": "#/components/schemas/UpdateUserRolesRelationInputDto"
          },
          "sessions": {
            "$ref": "#/components/schemas/UpdateUserSessionsRelationInputDto"
          }
        }
      },
...
      "ConnectUserDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "userId"
        ]
      },
...
      "UpdateRoleUsersRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectUserRoleDto"
            }
          }
        },
        "required": [
          "connect"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "users": {
            "$ref": "#/components/schemas/UpdateRoleUsersRelationInputDto"
          }
        }
      },
      "UserRoleQueryResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRole"
            }
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "data"
        ]
      },
      "CreateUserRoleUserRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "$ref": "#/components/schemas/ConnectUserDto"
          }
        },
        "required": [
          "connect"
        ]
      },
      "ConnectRoleDto": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string"
          }
        },
        "required": [
          "roleId"
        ]
      },
      "CreateUserRoleRoleRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "$ref": "#/components/schemas/ConnectRoleDto"
          }
        },
        "required": [
          "connect"
        ]
      },
      "CreateUserRoleDto": {
        "type": "object",
        "properties": {
          "userRoleId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/CreateUserRoleUserRelationInputDto"
          },
          "role": {
            "$ref": "#/components/schemas/CreateUserRoleRoleRelationInputDto"
          }
        },
        "required": [
          "user",
          "role"
        ]
      },
      "UpdateUserRoleUserRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "$ref": "#/components/schemas/ConnectUserDto"
          }
        },
        "required": [
          "connect"
        ]
      },
      "UpdateUserRoleRoleRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "$ref": "#/components/schemas/ConnectRoleDto"
          }
        },
        "required": [
          "connect"
        ]
      },
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UpdateUserRoleUserRelationInputDto"
          },
          "role": {
            "$ref": "#/components/schemas/UpdateUserRoleRoleRelationInputDto"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "access-token": []
    }
  ]
}